Save local changes before merge

+++ ray_tracing/RayTrace.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <cstdlib>
 #include <algorithm>
 #include <array>
@@ -88,7 +89,7 @@ struct Intersection {
 vector<Triangle*> constructSceneTriangles(vector<vector<int>> vertexBuffer, vector<int> indexBuffer, vector<vector<int>> normalBuffer){
     vector<Triangle*> sceneTriangles;
 
-    for(int i = 0; i < indexBuffer.size()-2; i+=3){
+    for(size_t i = 0; i < indexBuffer.size()-2; i+=3){
         //cout <<"a index: " << i  << " iB size: " << indexBuffer.size() << " iB value:" << indexBuffer[i] << " vB size: " << vertexBuffer.size() << endl;
         Point3 v1(vertexBuffer[indexBuffer[i]][0], vertexBuffer[indexBuffer[i]][1], vertexBuffer[indexBuffer[i]][2]);
         Point3 v2(vertexBuffer[indexBuffer[i+1]][0], vertexBuffer[indexBuffer[i+1]][1], vertexBuffer[indexBuffer[i+1]][2]);
@@ -227,7 +228,7 @@ Intersection* findIntersectingTriangle(Vector3 ray, vector<Triangle*> sceneTrian
     Triangle* tri = nullptr;
     Point3 intersection(0, 0, 0);
 
-    for(int i = 0; i < sceneTriangles.size(); i++){
+    for(size_t i = 0; i < sceneTriangles.size(); i++){
         HitInfo hit = mollerTrumbore(&ray, sceneTriangles[i]);
         // if(hit.hit){
         //     cout << "TRUE" << endl;
